adding post card component
